Some SPLToken instructions implemented
